+++ modules/features/player/src/main/java/au/com/shiftyjelly/pocketcasts/player/view/PlayerBottomSheet.kt
@@ -131,8 +131,8 @@ class PlayerBottomSheet @JvmOverloads constructor(context: Context, attrs: Attri
         binding.miniPlayer.setPlaybackState(playbackState)
     }
 
-    fun setUpNext(upNext: UpNextQueue.State, theme: Theme, shouldAnimateOnAttach: Boolean, useRssArtwork: Boolean) {
-        binding.miniPlayer.setUpNext(upNext, theme, useRssArtwork)
+    fun setUpNext(upNext: UpNextQueue.State, theme: Theme, shouldAnimateOnAttach: Boolean, useEpisodeArtwork: Boolean) {
+        binding.miniPlayer.setUpNext(upNext, theme, useEpisodeArtwork)
 
         // only show the mini player when an episode is loaded
         if (upNext is UpNextQueue.State.Loaded) {