+++ modules/features/player/src/main/java/au/com/shiftyjelly/pocketcasts/player/view/UpNextEpisodeViewHolder.kt
@@ -118,7 +118,7 @@ class UpNextEpisodeViewHolder(
             false
         }
 
-        imageRequestFactory.create(episode, settings.useRssArtwork.value).loadInto(binding.image)
+        imageRequestFactory.create(episode, settings.useEpisodeArtwork.value).loadInto(binding.image)
 
         val context = binding.itemContainer.context
         val transition = AutoTransition()