+++ modules/features/settings/src/main/java/au/com/shiftyjelly/pocketcasts/settings/viewmodel/SettingsAppearanceViewModel.kt
@@ -29,8 +29,7 @@ class SettingsAppearanceViewModel @Inject constructor(
     val signInState: LiveData<SignInState> = userManager.getSignInState().toLiveData()
     val createAccountState = MutableLiveData<SettingsAppearanceState>().apply { value = SettingsAppearanceState.Empty }
     val showArtworkOnLockScreen = settings.showArtworkOnLockScreen.flow
-    val useEmbeddedArtwork = settings.useEmbeddedArtwork.flow
-    val useRssArtwork = settings.useRssArtwork.flow
+    val useEpisodeArtwork = settings.useEpisodeArtwork.flow
 
     var changeThemeType: Pair<Theme.ThemeType?, Theme.ThemeType?> = Pair(null, null)
     var changeAppIconType: Pair<AppIcon.AppIconType?, AppIcon.AppIconType?> = Pair(null, null)
@@ -132,16 +131,8 @@ class SettingsAppearanceViewModel @Inject constructor(
         )
     }
 
-    fun updateUseEmbeddedArtwork(value: Boolean) {
-        settings.useEmbeddedArtwork.set(value, needsSync = true)
-        analyticsTracker.track(
-            AnalyticsEvent.SETTINGS_APPEARANCE_USE_EMBEDDED_ARTWORK_TOGGLED,
-            mapOf("enabled" to value),
-        )
-    }
-
-    fun updateUseRssArtwork(value: Boolean) {
-        settings.useRssArtwork.set(value, needsSync = true)
+    fun updateUseEpisodeArtwork(value: Boolean) {
+        settings.useEpisodeArtwork.set(value, needsSync = true)
         analyticsTracker.track(
             AnalyticsEvent.SETTINGS_APPEARANCE_USE_EPISODE_ARTWORK_TOGGLED,
             mapOf("enabled" to value),